GUI to test the VoiceOver thing.
Once you run the main method on this class, you are shown a frame with a
single button which opens up the JFileChooser dialogue. You can select
any file/directory and hit the "Read Name" button. This will display the
name of file/directory in the console.

VoiceOver does not work at this stage.