New GUI for MainFrame with textbox and fileMenu
Changed the dimensions of the GUI to fit the text box, file menu, and an
area for buttons. Added a file menu with "open", "save" and "clear" menu
items. You can open and save text files by using the respected menu
items and clear the text in textArea by pressing "clear". The "Sample"
button is a temp button that will be replaced by other buttons in the
future. When the button is pressed, a text of "Sample Text" is placed in
a new line. This action can be used later on when other buttons such as
"delay" are added. I only changed the variable name in the TopAuthoring
class.